Hack around awful replication. Not very successfully.
So this turns out to be a totally pathological case for Couch
replication. Couch leaves a tombstone for every deleted document.
That's fine and necessary. Otherwise you'd have no way of knowing a
thing was deleted. What sucks is that the replication protocol has to
fetch every single one of these zombie documents individually. The URL
along is longer than the document, let alone all the other crufty
overhead that goes along with the whole HTTP round trip.

This means that even with a modest amount of delted records, cold start
takes a _ridiculously_ long time. It's even worse on mobile. Which is
the whole point of this thing.

So you can do server-side filtering of replication functions. This
commit implements such a function that takes an array of deletions you
actually care about and ignores all the others.

That... sort of works. Cold start is a lot faster. It now seems pretty
unreliable in general though. Replication just stops working right.
There is less network traffic, but there's still a really remarkable
amount of it.

I suspect I am done trying to polish this particular turd. This just
turns out to be a really bad use case for Pouch. That sucks, because it
totally should work better, and it would if the Couch replication
protocol was better.